refactor: improved code readability and documentation in various classes

+++ src/main/java/com/ace/ucv/model/Grade.java
@@ -2,17 +2,17 @@ package com.ace.ucv.model;
 
 public class Grade {
 
-    private final float grade;
+    private final int grade;
     private final int subjectId;
     private final int studentId;
 
-    public Grade(float grade, int subjectId, int studentId) {
+    public Grade(int grade, int subjectId, int studentId) {
         this.grade = grade;
         this.subjectId = subjectId;
         this.studentId = studentId;
     }
 
-    public float getGrade() {
+    public int getGrade() {
         return grade;
     }
 
@@ -23,4 +23,6 @@ public class Grade {
     public int getStudentId() {
         return studentId;
     }
+
+
 }
